requirements.txt
Pillow==5.4.1
plotly==3.7.1
scipy==1.2.1